backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
psutil
PyYAML
requests
